fix(typeahead): fix the firing order of events
This fixes 2 issues that were affecting the typeahead and typeahead title. First, clicking outside
of the typeahead when using it in the title would not cause it to return to the title view instead
the typeahead stayed visible. Second, clikcing on a typeahead item to select would not cause it to
become the selected item because of event firing order.